requirements.txt
python-telegram-bot==20.7
psycopg2-binary==2.9.9
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-wtf==1.2.1
wtforms==3.1.1
pandas==2.1.4
openpyxl==3.1.2
reportlab==4.0.7
smtplib
email-validator==2.1.0
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
psutil==5.9.6
apscheduler==3.10.4
cryptography==41.0.8
bcrypt==4.1.2
jinja2==3.1.2
markupsafe==2.1.3
werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
sqlalchemy==2.0.23
alembic==1.13.1
mako==1.3.0
typing-extensions==4.8.0
anyio==4.2.0
httpx==0.25.2
certifi==2023.11.17
h11==0.14.0
sniffio==1.3.0
idna==3.6
httpcore==1.0.2
charset-normalizer==3.3.2
urllib3==2.1.0
requests==2.31.0
numpy==1.26.2
pytz==2023.3
six==1.16.0
et-xmlfile==1.1.0
pillow==10.1.0
kombu==5.3.4
billiard==4.2.0
vine==5.1.0
tzdata==2023.3